Struct boolean_expression::BDD

pub struct BDD<T> where T: Clone + Debug + Eq + Ord + Hash {
    // some fields omitted
}

A BDD is a Binary Decision Diagram, an efficient way to represent a Boolean function in a canonical way. (It is actually a "Reduced Ordered Binary Decision Diagram", which gives it its canonicity assuming terminals are ordered consistently.)

A BDD is built up from terminals (free variables) and constants, combined with the logical combinators AND, OR, and NOT. It may be evaluated with certain terminal assignments.

The major advantage of a BDD is that its logical operations are performed, it will "self-simplify": i.e., taking the OR of And(a, b) and And(a, Not(b)) will produce a without any further simplification step. Furthermore, the BDDFunc representing this value is canonical: if two different expressions are produced within the same BDD and they both result in (simplify down to) a, then the BDDFunc values will be equal. The tradeoff is that logical operations may be expensive: they are linear in BDD size, but BDDs may have exponential size (relative to terminal count) in the worst case.

Methods

impl<T> BDD<T> where T: Clone + Debug + Eq + Ord + Hash

fn new() -> BDD<T>

Produce a new, empty, BDD.

fn terminal(&mut self, t: T) -> BDDFunc

Produce a function within the BDD representing the terminal t. If this terminal has been used in the BDD before, the same BDDFunc will be returned.

fn constant(&mut self, val: bool) -> BDDFunc

Produce a function within the BDD representing the constant value val.

fn not(&mut self, n: BDDFunc) -> BDDFunc

Produce a function within the BDD representing the logical complement of the function n.

fn and(&mut self, a: BDDFunc, b: BDDFunc) -> BDDFunc

Produce a function within the BDD representing the logical AND of the functions a and b.

fn or(&mut self, a: BDDFunc, b: BDDFunc) -> BDDFunc

Produce a function within the BDD representing the logical OR of the functions a and b.

fn expr(&mut self, e: &Expr<T>) -> BDDFunc

Produce a function within the BDD representing the given expression e, which may contain ANDs, ORs, NOTs, terminals, and constants.

fn evaluate(&self, f: BDDFunc, values: &HashMap<T, bool>) -> bool

Evaluate the function f in the BDD with the given terminal assignments. Any terminals not specified in values default to false.