Struct boolean_expression::Cube

pub struct Cube(_);

A Cube is one (multidimensional) cube in the variable space: i.e., a particular set of variable assignments, where each variable is assigned either true, false, or "don't care".

Methods

impl Cube

fn true_cube(vars: usize) -> Cube

Construct an always-true cube (all variables are DontCare) for vars variables.

fn vars(&self) -> Iter<CubeVar>

Return an iterator over variable assignments.

fn merge_with(&self, other: &Cube) -> CubeMergeResult

Attempt to merge this cube with another, which may cancel one or the other (if completely covered), expand one or the other (if possible, to increase the number of DontCares and thus simplify the final expression), or merge the two into a single cube, or do nothing.

fn with_var(&self, idx: usize, val: CubeVar) -> Cube

Return a new cube equal to this cube, but with the particular variable