Struct boolean_expression::CubeList

pub struct CubeList(_);

A CubeList is a sum (OR'd list) of cubes. It represents a Boolean expression in sum-of-products form, by construction.

The CubeList abstraction supports only indexed anonymous variables (variable 0, 1, ...), and does not (yet) have a wrapper supporting an arbitrary terminal type T. This may be implemented in the future.

The CubeList abstraction is used internally to convert from a BDD to a quasi-minimized Boolean expression.

Methods

impl CubeList

fn new() -> CubeList

Construct a new, empty, cube list (equivalent to a constant false).

fn from_list(cubes: &[Cube]) -> CubeList

Construct a cube list from a list of Cube structs.

fn cubes(&self) -> Iter<Cube>

Return an iterator over all cubes.

fn merge(&self, other: &CubeList) -> CubeList

Merge this cube list with another, canceling or merging cubes where possible. The resulting cube list is not guaranteed to be minimal (that is the set-cover problem, which is NP-Complete), but is reduced somewhat by a very simple reduction/merging algorithm.

fn with_var(&self, idx: usize, val: CubeVar) -> CubeList